blocks sc_protect_funcall
the "random" crash bug was there because rb_funcall was being called
and it was not blocked, so when a scheduled method (with a really small
frequency) tried to call a ruby method, some race condition was met.
The best approach is to block all ruby calls from the Obj-C world, that
means, blocking in sc_protect_funcall, which is the only way ShinyCocos
calls ruby methods.